libbass: 2.4.15 -> 2.4.16
Upstream only offers a download link to the latest version, so to keep
this package working we have no choice but to keep it always up-to-date.

In a previous discussion[1] it was suggested that we keep snapshots of
old versions, but it seems this was never done. Right now it shouldn't
be a problem as all 2.4.x releases should be compatible with eachother.

[1]: http://www.un4seen.com/forum/?topic=18614.0